requirements.txt
django==3.0.7
djangorestframework==3.11.0
django-cors-headers==3.4.0
drf-flex-fields==0.8.5
django-filter